+++ lihil/plugins/jwt.py
@@ -1,8 +1,0 @@
-# TODO: builtin-jwt mechanism
-
-
-from lihil.interface import Payload, Form
-
-
-
-# type LoginForm = Form[]